Move nesting checks in common module
